requirements.txt
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Flask==3.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
paramiko==3.5.0
psycopg2-binary==2.9.10
pycparser==2.22
python-dotenv==1.0.0
PyNaCl==1.5.0
python-gnupg==0.5.2
requests==2.31.0
resend==0.7.2
urllib3==2.2.3
Werkzeug==3.1.3
